backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
rembg[cpu]>=2.0.0
pillow>=10.0.0
